intとunsignedの比較警告がわずらわしいので明示キャストしました。
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@5574 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Creature/CRBody.h
@@ -59,7 +59,7 @@ public:
 	/** @brief i”Ô–Ú‚Ì„‘Ì‚ğ“¾‚é
 	*/
 	virtual CRSolidIf* GetSolid(int i) {
-		if (0<=i && i<solids.size()) {
+		if (0<=i && i<(int)solids.size()) {
 			return solids[i];
 		} else {
 			return NULL;
@@ -73,7 +73,7 @@ public:
 	/** @brief i”Ô–Ú‚ÌŠÖß‚ğ“¾‚é
 	*/
 	virtual CRJointIf* GetJoint(int i) {
-		if (0<=i && i<joints.size()) {
+		if (0<=i && i<(int)joints.size()) {
 			return joints[i];
 		} else {
 			return NULL;
